@ranocha 2024-08-18
Simulations could be much faster if https://github.com/SciML/RecursiveArrayTools.jl/issues/400 was resolved...