@ranocha 2024-08-18
@JoshuaLampert Do you know why you set the lower compat of RecursiveArrayTools.jl to 0.3.3 in https://github.com/JoshuaLampert/DispersiveShallowWater.jl/pull/118?