@JoshuaLampert 2024-08-18
Thanks a lot! The performance improvement already looks quite nice.

> To get values that are closer to our paper, we have to omit the analysis callback (which allocates quite a lot):

Can you appraise whether the `AnalysisCallback` needs to allocate so much or do you think we can optimize it (not necessarily in this PR)?

> However, this is still significantly slower (between 5% and 25%).

Do you have an idea where we loose the remaining 5 to 25%?

Do you recommend using FastBroadcast.jl for the other equations, too? Probably the performance gain would not be too high because the solving the elliptic equations is the most expensive, but I take what I can get. If you think it could help, I could do that in a future PR.

> @JoshuaLampert Do you know why you set the lower compat of RecursiveArrayTools.jl to 0.3.3 in https://github.com/JoshuaLampert/DispersiveShallowWater.jl/pull/118? This is involved in compatibility issues on Julia 1.9...

I don't remember in detail, but judging from the commit history in that PR and the respective CI runs, it looks like it was because [RecursiveArrays.jl >v0.3.3 doesn't support Julia v1.9](https://github.com/JoshuaLampert/DispersiveShallowWater.jl/actions/runs/9550667338/job/26323186641#step:6:55). This comes from https://github.com/SciML/RecursiveArrayTools.jl/pull/324, which was included in [v0.3.4 of RecursiveArrays.jl](https://github.com/SciML/RecursiveArrayTools.jl/releases/tag/v3.4.0).