@ranocha 2024-08-19
> > To get values that are closer to our paper, we have to omit the analysis callback (which allocates quite a lot):
> 
> Can you appraise whether the `AnalysisCallback` needs to allocate so much or do you think we can optimize it (not necessarily in this PR)?

At least the modified energy/entropy allocates a full space vector every time it is called