@ranocha 2024-08-19
> Do you have an idea where we loose the remaining 5 to 25%?

We compute the derivative of the bathymetry `b_x` from scratch every time. Moreover, we have some additional overhead from storing `eta` instead of `h` and `D` instead of `b` etc. Maybe we could get rid of some of the overhead, but we would have to analyze the split form in terms of the variables we use in this repo