@ranocha 2024-08-19
> Do you recommend using FastBroadcast.jl for the other equations, too? Probably the performance gain would not be too high because the solving the elliptic equations is the most expensive, but I take what I can get. If you think it could help, I could do that in a future PR.

As you said, I expect the elliptic solve(s) to dominate significantly for the other equations. Maybe it could help a bit but I would not expect significant performance gains liek here